requirements.txt

evol
deap~=1.3.3
Pillow~=8.4.0